Anime and Settings XML Refactor; New Classes
A general rework on how settings and anime properties are retrieved, set on any property will save to the xml document and much less work should need to be done through the controlling XML classes with the work being done more 'behind the scenes' when accessing and setting values.

To accomodate the change in how to access those classes and to be maybe more modular are the introduction to a few sets of XML related classes. General refactoring was done around these changes as well as some refactoring from basic static analysis.