A shift in design to how views interact with eachother
Instead of implementing the details of how all views work inside MainWindow through applying delegate functions to components, MainWindow was redesigned to work as the holder of all "global" variables and able to be accessed statically for it's contents and have each view call this window to access needed variables. This also gives the benefit that I can implement more functionality in the XAML because I can access the global objects within the other views, and so MainWindow is mostly cleared of functions that are meant to influence other views. This heavily condenses some sections of code and makes some sections look overall cleaner, and allowed me to decouple functions to condense them and move them into more appropriate locations

This change refactors a lot of code to new locations.

Other changes:
1. Fix settings to not implicitly autosave
2. Clicking header labels in manage will open those folders
3. Update README slightly
4. New "Enums" category, with future use of anime status enums to be determined
5. Helper methods in the anime class IntRating && HasRating
6. Move other distance method to correct location
7. Small class name changes
8. Move the slightly complex FindBox functionality in AnimeList to a new class
9. Tray menu selections have shortcut keys (Though I can't figure out how to display them in the text)