Removed finding ID within animedetails for now
This caused a long standing issue that I couldn't find the issue with. Apparently on MyAnimeList you can "update" an anime that you haven't added yet so while the update would be successfull showing all details like rating and episode watched, since it was never added it never appeared on the list.