Use ViewModel IOC for services, Tray impl, see details for more
To integrate the viewmodels into the IOC container, I had to move the tray to the mainwindowviewmodel so this was kind of a two for one deal. Small other changes include: Remove pressing enter on fields in place of IsDefault, reinstantiation of view models with messageinstance listeners will have explicit Cleanup() calls in their property {get;set;} to avoid have them lingering in memory and still listening for messages.