Create a Playlist class to soon replace the playlist service
Using a service sort of constraints having multiple instances of playlists, which I found out when I wanted to have a 'demo' playlist showing manipulation of the sorts but was unable. This implementation will allow me to do so more fluidly without needing to inject it into any service or grab it from the ioc container.