Create base class for webservices, abstract nyaaservice further
DownloadService implementations will generally have a set of core functionality that will be identical between them such as starting the download from a given url or attempting download from an enumerable of torrents. DownloadServiceBase as the base abstract inheritor of webservices allows implementations (like NyaaService right now) to only need to specify finding the list of torrents and service name while still maintaining the expected functionality.

In the future this might become muddier, as StartTorrent or DownloadTorrent in the interface and base class rely on the idea of 'Torrent' where in the future it might need to be tiered backwards to accomodate for other methods of transfer (IRC client, direct download, ...)